TRUS-2676 Correct typos on renaming views and controllers
