+++ app/controllers/print/PrintMaintainDraftAnswersController.scala
@@ -22,7 +22,7 @@ import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import utils.print.PrintPlaybackHelper
-import views.html.print.PlaybackDraftAnswersView
+import views.html.print.PrintMaintainDraftAnswersView
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -30,7 +30,7 @@ class PrintMaintainDraftAnswersController @Inject()(
                                                 override val messagesApi: MessagesApi,
                                                 actions: AuthenticateForPlayback,
                                                 val controllerComponents: MessagesControllerComponents,
-                                                view: PlaybackDraftAnswersView,
+                                                view: PrintMaintainDraftAnswersView,
                                                 printPlaybackAnswersHelper: PrintPlaybackHelper
                                               )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
