+++ conf/messages.en
@@ -673,11 +673,15 @@ taskList.in-progress = In progress
 
 # Playback prints
 
+# Last declared answers
+
 playbackAnswers.title = Last declared copy of the trust’s registration
 playbackAnswers.heading = Last declared copy of the trust’s registration
 playbackAnswers.declareAnnually.warning = You only need to declare the trust is up to date every year if there is a tax liability.
 playbackAnswers.informationFirstRegistered = The following sections indicate what was entered when the trust was first registered. No further updates to trust details, assets and years of tax liability are required through this service.
 playbackAnswers.selfAssessmentOnline = If you need to update these sections, use Self Assessment for trusts.
+
+# Declared answers following maintaining the trust
 
 playbackDeclarationAnswers.title = Declared copy of updates to the trust
 playbackDeclarationAnswers.heading = Declared copy of updates to the trust
@@ -686,6 +690,8 @@ playbackDeclarationAnswers.p2 = Client reference number: {0}
 playbackDeclarationAnswers.p3 = The trust’s declaration was sent on {0}
 playbackDeclarationAnswers.warning = You only need to declare the trust is up to date every year if there is a tax liability.
 
+# Draft answers after updating and maintaining the trust
+
 playbackDraftAnswers.title = Draft copy of the trust's declaration
 playbackDraftAnswers.heading = Draft copy of the trust's declaration
 playbackDraftAnswers.p1 = This is a draft copy confirming that the trust is up to date.
@@ -693,6 +699,8 @@ playbackDraftAnswers.p2 = If you have made any changes they must reflect what's 
 playbackDraftAnswers.p3 = Any details shown for assets and years of tax liability indicate what was entered at registration. No further updates to these details are required through this service.
 playbackDraftAnswers.p4 = This is not a declaration.
 playbackDraftAnswers.selfAssessmentOnline = If you need to view the latest information HMRC holds about these sections or need to update them, use Self Assessment Online for trusts.
+
+# Final declaration following closing the trust
 
 playbackFinalDeclarationAnswers.title = Final declared copy of updates to this trust
 playbackFinalDeclarationAnswers.heading = Final declared copy of updates to this trust
